+++ src/Contracts/MarkdownFileHandler.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Cassarco\MarkdownTools\Contracts;
+
+use Cassarco\MarkdownTools\MarkdownFile;
+
+interface MarkdownFileHandler
+{
+    public function __invoke(MarkdownFile $file): void;
+}