+++ src/Contracts/MarkdownFileRules.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Cassarco\MarkdownTools\Contracts;
+
+interface MarkdownFileRules
+{
+    public function __invoke(): array;
+}