+++ src/Exceptions/InvalidConfigException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Cassarco\MarkdownTools\Exceptions;
+
+use Exception;
+
+class InvalidConfigException extends Exception
+{
+}